KOGITO-5401 - Remove Kogito Quarkus archetype
