--- archetypes/.gitignore
@@ -1,14 +1,0 @@
-/target
-/local
-
-# Eclipse, Netbeans and IntelliJ files
-/.*
-!.gitignore
-/nbproject
-/*.ipr
-/*.iws
-/*.iml
-*.xpr
-
-# Repository wide ignore mac DS_Store files
-.DS_Store
