--- archetypes/kogito-quarkus-archetype/src/main/resources/archetype-resources/src/test/java/GreetingsTest.java
@@ -1,44 +1,0 @@
-/*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package ${package};
-
-import io.quarkus.test.junit.QuarkusTest;
-import io.restassured.http.ContentType;
-import org.junit.jupiter.api.Test;
-
-import java.util.UUID;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.CoreMatchers.*;
-
-@QuarkusTest
-public class GreetingsTest {
-
-    @Test
-    public void testHelloEndpoint() {
-        given()
-            .contentType(ContentType.JSON)
-            .accept(ContentType.JSON)
-            .body("{}")
-        .when()
-            .post("/greetings")
-        .then()
-            .statusCode(201)
-            .header("Location", notNullValue())
-            .body("id", notNullValue());
-    }
-
-}
