--- archetypes/kogito-quarkus-dm-archetype/src/main/resources/archetype-resources/src/test/resources/application.properties
@@ -1,5 +1,0 @@
-# Quarkus
-quarkus.http.test-port=0
-quarkus.swagger-ui.always-include=true
-
-kogito.service.url=http://localhost:8080
