--- archetypes/kogito-quarkus-dm-archetype/src/test/resources/projects/it-basic/archetype.properties
@@ -1,4 +1,0 @@
-groupId=it.pkg
-artifactId=basic-project
-version=0.1-SNAPSHOT
-package=it.pkg
