--- archetypes/kogito-quarkus-dm-archetype/src/test/resources/projects/it-basic/goal.txt
@@ -1,1 +1,0 @@
-verify -Dversion.io.quarkus=${version.io.quarkus}
