--- archetypes/kogito-springboot-archetype/src/main/resources/archetype-resources/src/main/resources/test-process.bpmn2
@@ -1,99 +1,0 @@
-<bpmn2:definitions xmlns:bpmn2="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:bpsim="http://www.bpsim.org/schemas/1.0" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" xmlns:drools="http://www.jboss.org/drools" id="_wS31gVJREDiOzfiIY5TGdQ" exporter="jBPM Process Modeler" exporterVersion="2.0" targetNamespace="http://www.omg.org/bpmn20">
-  <bpmn2:process id="greetings" drools:packageName="\${groupId}" drools:version="1.0" drools:adHoc="false" name="greetings" isExecutable="true" processType="Public">
-    <bpmn2:sequenceFlow id="_D764D3B4-1533-4962-A9E3-739A6DD9AEAB" sourceRef="_3CDC6E61-DCC5-4831-8BBB-417CFF517CB0" targetRef="_A6902151-5E9D-48F7-95E9-375E41CF3E6F">
-      <bpmn2:extensionElements>
-        <drools:metaData name="isAutoConnection.source">
-          <drools:metaValue>true</drools:metaValue>
-        </drools:metaData>
-        <drools:metaData name="isAutoConnection.target">
-          <drools:metaValue>true</drools:metaValue>
-        </drools:metaData>
-      </bpmn2:extensionElements>
-    </bpmn2:sequenceFlow>
-    <bpmn2:sequenceFlow id="_00AB4A77-D70F-4086-8BA6-57DD017A5323" sourceRef="_75AC8C0C-CFBD-4ADF-A3B4-83AB90581A73" targetRef="_3CDC6E61-DCC5-4831-8BBB-417CFF517CB0">
-      <bpmn2:extensionElements>
-        <drools:metaData name="isAutoConnection.source">
-          <drools:metaValue>true</drools:metaValue>
-        </drools:metaData>
-        <drools:metaData name="isAutoConnection.target">
-          <drools:metaValue>true</drools:metaValue>
-        </drools:metaData>
-      </bpmn2:extensionElements>
-    </bpmn2:sequenceFlow>
-    <bpmn2:endEvent id="_A6902151-5E9D-48F7-95E9-375E41CF3E6F">
-      <bpmn2:incoming>_D764D3B4-1533-4962-A9E3-739A6DD9AEAB</bpmn2:incoming>
-    </bpmn2:endEvent>
-    <bpmn2:scriptTask id="_3CDC6E61-DCC5-4831-8BBB-417CFF517CB0" name="Hello" scriptFormat="http://www.java.com/java">
-      <bpmn2:extensionElements>
-        <drools:metaData name="elementname">
-          <drools:metaValue>Hello</drools:metaValue>
-        </drools:metaData>
-      </bpmn2:extensionElements>
-      <bpmn2:incoming>_00AB4A77-D70F-4086-8BA6-57DD017A5323</bpmn2:incoming>
-      <bpmn2:outgoing>_D764D3B4-1533-4962-A9E3-739A6DD9AEAB</bpmn2:outgoing>
-      <bpmn2:script>System.out.println("Hello World");</bpmn2:script>
-    </bpmn2:scriptTask>
-    <bpmn2:startEvent id="_75AC8C0C-CFBD-4ADF-A3B4-83AB90581A73">
-      <bpmn2:outgoing>_00AB4A77-D70F-4086-8BA6-57DD017A5323</bpmn2:outgoing>
-    </bpmn2:startEvent>
-  </bpmn2:process>
-  <bpmndi:BPMNDiagram>
-    <bpmndi:BPMNPlane bpmnElement="greetings">
-      <bpmndi:BPMNShape id="shape__75AC8C0C-CFBD-4ADF-A3B4-83AB90581A73" bpmnElement="_75AC8C0C-CFBD-4ADF-A3B4-83AB90581A73">
-        <dc:Bounds height="56" width="56" x="176" y="198"/>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="shape__3CDC6E61-DCC5-4831-8BBB-417CFF517CB0" bpmnElement="_3CDC6E61-DCC5-4831-8BBB-417CFF517CB0">
-        <dc:Bounds height="76" width="143" x="315" y="188"/>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="shape__A6902151-5E9D-48F7-95E9-375E41CF3E6F" bpmnElement="_A6902151-5E9D-48F7-95E9-375E41CF3E6F">
-        <dc:Bounds height="56" width="56" x="538" y="198"/>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNEdge id="edge_shape__75AC8C0C-CFBD-4ADF-A3B4-83AB90581A73_to_shape__3CDC6E61-DCC5-4831-8BBB-417CFF517CB0" bpmnElement="_00AB4A77-D70F-4086-8BA6-57DD017A5323">
-        <di:waypoint x="232" y="226"/>
-        <di:waypoint x="315" y="226"/>
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNEdge id="edge_shape__3CDC6E61-DCC5-4831-8BBB-417CFF517CB0_to_shape__A6902151-5E9D-48F7-95E9-375E41CF3E6F" bpmnElement="_D764D3B4-1533-4962-A9E3-739A6DD9AEAB">
-        <di:waypoint x="458" y="226"/>
-        <di:waypoint x="538" y="226"/>
-      </bpmndi:BPMNEdge>
-    </bpmndi:BPMNPlane>
-  </bpmndi:BPMNDiagram>
-  <bpmn2:relationship id="_wTJiUFJREDiOzfiIY5TGdQ" type="BPSimData">
-    <bpmn2:extensionElements>
-      <bpsim:BPSimData>
-        <bpsim:Scenario id="default" name="Simulationscenario">
-          <bpsim:ScenarioParameters/>
-          <bpsim:ElementParameters elementRef="_75AC8C0C-CFBD-4ADF-A3B4-83AB90581A73">
-            <bpsim:TimeParameters>
-              <bpsim:ProcessingTime>
-                <bpsim:NormalDistribution mean="0" standardDeviation="0"/>
-              </bpsim:ProcessingTime>
-            </bpsim:TimeParameters>
-          </bpsim:ElementParameters>
-          <bpsim:ElementParameters elementRef="_3CDC6E61-DCC5-4831-8BBB-417CFF517CB0">
-            <bpsim:TimeParameters>
-              <bpsim:ProcessingTime>
-                <bpsim:NormalDistribution mean="0" standardDeviation="0"/>
-              </bpsim:ProcessingTime>
-            </bpsim:TimeParameters>
-            <bpsim:ResourceParameters>
-              <bpsim:Availability>
-                <bpsim:FloatingParameter value="0"/>
-              </bpsim:Availability>
-              <bpsim:Quantity>
-                <bpsim:FloatingParameter value="0"/>
-              </bpsim:Quantity>
-            </bpsim:ResourceParameters>
-            <bpsim:CostParameters>
-              <bpsim:UnitCost>
-                <bpsim:FloatingParameter value="0"/>
-              </bpsim:UnitCost>
-            </bpsim:CostParameters>
-          </bpsim:ElementParameters>
-        </bpsim:Scenario>
-      </bpsim:BPSimData>
-    </bpmn2:extensionElements>
-    <bpmn2:source>_wS31gVJREDiOzfiIY5TGdQ</bpmn2:source>
-    <bpmn2:target>_wS31gVJREDiOzfiIY5TGdQ</bpmn2:target>
-  </bpmn2:relationship>
-</bpmn2:definitions>
