--- archetypes/pom.xml
@@ -1,68 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-build-parent</artifactId>
-    <version>2.0.0-SNAPSHOT</version>
-    <relativePath>../kogito-build/kogito-build-parent/pom.xml</relativePath>
-  </parent>
-  <artifactId>archetypes</artifactId>
-  <packaging>pom</packaging>
-  <name>Kogito :: Maven Archetypes</name>
-  <description>Various Kogito Maven archetypes for project generation</description>
-
-  <properties>
-    <maven.compiler.release>11</maven.compiler.release> 
-  </properties>
-
-  <build>
-    <pluginManagement>
-      <plugins>
-        <plugin>
-          <artifactId>maven-archetype-plugin</artifactId>
-          <configuration>
-            <localRepositoryPath>${session.request.localRepositoryPath.path}</localRepositoryPath>
-          </configuration>
-        </plugin>
-      </plugins>
-    </pluginManagement>
-  </build>
-
-  <profiles>
-    <profile>
-      <id>default</id>
-      <activation>
-        <property>
-          <name>!productized</name>
-        </property>
-      </activation>
-      <modules>
-        <module>kogito-springboot-archetype</module>
-        <module>kogito-quarkus-archetype</module>
-      </modules>
-    </profile>
-    <profile>
-      <!--
-        This profile serves at optionally passing settingsFile property to archetype::integration-test goal.
-        Reason is that the ${session.request.userSettingsFile.path} property defaults to ~/.m2/settings.xml even
-        if that file does not actually exist. Tests were in such case failing.
-      -->
-      <id>set-settings-when-provided</id>
-      <activation>
-        <file><exists>${session.request.userSettingsFile.path}</exists></file>
-      </activation>
-      <build>
-        <pluginManagement>
-          <plugins>
-            <plugin>
-              <artifactId>maven-archetype-plugin</artifactId>
-              <configuration>
-                <settingsFile>${session.request.userSettingsFile.path}</settingsFile>
-              </configuration>
-            </plugin>
-          </plugins>
-        </pluginManagement>
-      </build>
-    </profile>
-  </profiles>
-</project>
