+++ springboot/archetype/src/main/resources/META-INF/archetype-post-generate.groovy
