+++ springboot/archetype/src/main/resources/META-INF/maven/archetype-metadata.xml
