+++ springboot/archetype/src/main/resources/archetype-resources/src/main/java/CorsConfig.java
