+++ springboot/archetype/src/main/resources/archetype-resources/src/main/java/KogitoApplication.java
