+++ springboot/archetype/src/main/resources/archetype-resources/src/main/resources/META-INF/resources/index.html
