+++ springboot/archetype/src/main/resources/archetype-resources/src/main/resources/TrafficViolation.dmn
