+++ springboot/archetype/src/main/resources/archetype-resources/src/main/resources/application.properties
