+++ springboot/archetype/src/main/resources/archetype-resources/src/main/resources/test-process.bpmn2
