+++ springboot/archetype/src/main/resources/archetype-resources/src/test/java/GreetingsTest.java
