+++ springboot/archetype/src/test/resources/projects/it-basic/archetype.properties
