+++ springboot/archetype/src/test/resources/projects/it-basic/goal.txt
