+++ springboot/archetype/src/test/resources/projects/it-custom-decisions/archetype.properties
