+++ springboot/archetype/src/test/resources/projects/it-custom-decisions/goal.txt
