+++ springboot/archetype/src/test/resources/projects/it-custom-process/goal.txt
